+++ Core/Src/main.c
@@ -353,7 +353,7 @@ int main(void)
   start_Si5351();
 
   Set_Cursor_Frequency();
-  show_variable(400, 25, (int)NCO_Frequency);
+  show_variable(400, 25, NCO_Frequency);
   show_short(667, 255, AGC_Gain);
   start_duplex();
   HAL_Delay(10);
